Refactor backend to replace `ast` structures with `tacky` representation and enhance assembler and codegen pipelines. Adjust frontend parser for compatibility.

+++ src/frontend/ast.rs
@@ -46,7 +46,7 @@ impl Stmt {
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum ExprKind {
-    Constant(i64),
+    Constant(i32),
     Unary(UnaryOp, Box<Expr>),
     Return(Option<Box<Expr>>),
 }
